Create sysinfo.sh
This is supposed to run and create of the files needed, however I ran into big issues actually getting something to test it on my laptop, I even tried the google cloud terminal and that did not work. So this is an untested commit that I just have to submit so yeah.