Finished Commit
This commit I have finished all of the code and it all worked for me, apart from that on weird issue that was discussed with the TA. Other than that everything should work fine, there is one weird issue where if the file does no have one extra blank line, then the code will add the extra username to the last one on the list. Other than that it appears to work.